Fix an 32 bit truncation on linkRep.

+++ models/starComplexity.h
@@ -1,4 +1,4 @@
-using linkRep=unsigned;
+using linkRep=unsigned long long;
 
 #include <map>
 #include <unordered_map>
